fix filter bug
